Don't delete the "diff" and "index" headers
Emacs 23.4 pretest has changes to diff-mode such that it no longer
matches Git's "---" and "+++" header lines correctly, and without the
"diff" header line generated by Git, commands like diff-goto-source
(bound to RET in gitsum) can't find the file that a hunk belongs to.
Let's leave the "diff" and "index" headers in so diff-goto-source
keeps working.  I hopefully made gitsum-kill-dwim behave properly when
you use it when on a "diff" or "index" header.